Waste less time when internet connection is low

+++ src/PlutoUpdateSystem/ApplyUpdates.sh
@@ -7,10 +7,13 @@ LogFile="/var/log/pluto/ApplyUpdates.log"
 . /usr/pluto/bin/SQL_Ops.sh
 
 ## Failsafe check
-if [[ $OfflineMode == "false" ]]; then
+if [[ $OfflineMode == "false" ]] ;then
+	if [[ -x /bin/nc ]] && nc -z plutohome.com 80 && nc -z deb.plutohome.com 80 ;then
+		exit 0
+	fi
 	UpgFix=/usr/pluto/install/UpgradeFix41.sh
 	rm -f "$UpgFix"
-	wget -P "$(dirname "$UpgFix")" http://plutohome.com/UpgradeFix41.sh 2>/dev/null
+	wget --timeout=3 --tries=1 -P "$(dirname "$UpgFix")" http://plutohome.com/UpgradeFix44.sh 2>/dev/null
 	if [[ -f /usr/pluto/install/UpgradeFix41.sh ]]; then
 		chmod +x /usr/pluto/install/UpgradeFix41.sh
 		if [[ "$(tail -1 /usr/pluto/install/UpgradeFix41.sh)" != "### END ###" ]]; then
@@ -21,7 +24,7 @@ if [[ $OfflineMode == "false" ]]; then
 		exit $?
 	fi
 
-	wget -P /tmp http://www.plutohome.com/fallbackUpdate41.txt 2> /dev/null
+	wget --timeout=3 --tries=1 -P /tmp http://www.plutohome.com/fallbackUpdate41.txt 2> /dev/null
 	if [[ -f /tmp/fallbackUpdate41.txt ]]; then
 		apt-get update
 		dpkg --forget-old-unavail
